requirements.txt
streamlit==1.15.0
opencv-python-headless==4.7.0.72
numpy==1.24.0
tensorflow==2.15.0
Pillow==9.4.0
sqlite3==3.41.2
